requirements.txt
allpairspy==2.5.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
config==0.5.0.post0
crontab==0.23.0
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docx==0.2.4
idna==3.2
lxml==4.6.3
pandas==1.3.1
Pillow==8.3.1
PyMySQL==0.9.3
pypinyin==0.42.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
xmindparser==1.0.9
prettytable~=2.2.0
jsonpath~=0.82
pytest~=6.2.5
allure-pytest==2.9.43
pytest-dependency==0.5.1
pytest-html==3.1.1
pycryptodome==3.10.1
progressbar~=2.5
arrow~=1.1.1
Faker~=8.12.1
loguru==0.5.3
jsonpath==0.82
jmespath==0.10.0
pydantic~=1.8.2
Jinja2~=3.0.2
sentry-sdk~=1.4.3
dictdiffer~=0.9.0
pymongo~=3.12.1
py~=1.10.0
python-jenkins~=1.7.0
yapf~=0.31.0
djangorestframework-bulk~=0.2.1
dwebsocket~=0.5.12